rename ampelprepper to ampel interafec and move resukts stuff there
