+++ timewise/process/__init__.py
@@ -1,2 +1,2 @@
 from .config import AmpelConfig
-from .prepare import AmpelPrepper
+from .interface import AmpelInterface
