+++ timewise/process/interface.py
@@ -4,12 +4,13 @@ import logging
 import pandas as pd
 from pymongo import MongoClient, ASCENDING
 from ampel.cli.JobCommand import JobCommand
+from ampel.types import StockId
 
 
 logger = logging.getLogger(__name__)
 
 
-class AmpelPrepper:
+class AmpelInterface:
     def __init__(
         self,
         mongo_db_name: str,
@@ -78,3 +79,18 @@ class AmpelPrepper:
         )
         logger.debug(args)
         cmd.run(args, unknown_args=())
+
+    @property
+    def client(self) -> MongoClient:
+        return MongoClient(self.uri)
+
+    def extract_stacked_lightcurve(self, stock_id: StockId) -> pd.DataFrame:
+        col = self.client[self.mongo_db_name]["t1"]
+        records = []
+        for i, ic in enumerate(col.find({"stock": stock_id, "unit": "T1StackVisits"})):
+            assert i == 0, f"More than one stacked lightcurve found for {stock_id}!"
+            assert len(ic["body"]) == 1, (
+                f"None or more than one stacking result found for {stock_id}!"
+            )
+            records = ic["body"][0]
+        return pd.DataFrame(records)
